Move SearchRank enum to search package
